+++ changelog.md
@@ -3,7 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.1] - 2022-04-28
+
+## [2.0.4] - 2022-05-06
+
+- changed
+  - re-declaration of value removed
+- added
+  - test class instance case
+## [2.0.1] - deprecated - 2022-04-28
+
 - changed
   - package name: `proxy-handler-generic-execution` -> `generic-interceptor`
   - `ProxyHandlerGenericExecutionOptions` -> `InterceptorOptions`